refactor: split project card into reusable component
