+++ src/components/Project/Projects.json
@@ -1,11 +1,32 @@
 [
   {
     "id": 1,
+    "title": "Frontend - 푸딘코 테이블오더",
+    "tier": "primary",
+    "image": ["table1.png", "table2.png"],
+    "date": "2025.01 ~ 2026.02 / 2인 (FE 1, BE 1)",
+    "descript": "소켓 기반 실시간 주문 동기화를 지원하는 식당용 QR 테이블오더 서비스입니다. 로컬 환경 구현부터 개발 서버 테스트, 운영 배포까지 전 과정을 경험하며 실제 서비스 환경에서의 프론트엔드 역할과 책임을 체감했습니다.",
+    "skills": "PHP, JavaScript, jQuery, CSS, Socket.io, AWS EC2, Apache, Redis, MariaDB, Figma",
+    "response": "- 프론트엔드 단독 담당\n- 신규 기능 기획, 화면 설계, 개발 및 QA 전 과정 수행\n- 실서비스 런칭 과정에서 발생한 FE 이슈 대응 및 서비스 안정화 기여"
+  },
+  {
+    "id": 2,
+    "title": "Frontend - 푸딘코 리액트 웹앱",
+    "tier": "primary",
+    "image": ["react1.png", "react2.png"],
+    "date": "2025.10 ~ 2025.12 / 3인 (FE 2, BE 1)",
+    "descript": "지도 기반 탐색과 검색을 통해 제휴 매장을 쉽게 찾고, 매장의 상세 정보 및 공유 기능을 제공하는 식당 큐레이션 웹 서비스입니다. 프로젝트 초기부터 유지보수와 확장성을 고려해 디렉토리 구조와 상태 관리 방식을 설계하며 개발을 진행했습니다.",
+    "skills": "React, TypeScript, Vite, Zustand, React Query, Axios, Tailwind CSS, AWS EC2, Apache",
+    "response": "- 프론트엔드 선임 역할 수행\n- 프로젝트 초기 레포 구조 및 기술 스택 선정 주도\n- 아키텍처 설계 및 구현, 서버 상태와 클라이언트 상태 분리 구조 설계\n- 배포 환경에서 SPA 라우팅 이슈 분석 및 개선 경험"
+  },
+  {
+    "id": 3,
     "title": "Frontend - 초보 농부를 위한 정보 제공 사이트 '밭드림'",
+    "tier": "primary",
     "image": ["team3-0.png", "team3-1.png", "team3-2.png"],
     "date": "2024.04.26 ~ 05.16 / 6인",
     "descript": "가족이 먹을 농작물을 주말농장에서 직접 재배하는 인구는 늘고 있지만, 관련 지식이 없으면 접근이 어렵다는 점에서 착안하여 기획 및 제작한 사이트입니다. 농작물의 성장에는 날씨가 중요하여 날씨&미세먼지 위젯을 담당했습니다. React 프로젝트와 fetch()로 데이터를 출력하는 과정을 처음 경험해보며, 컴포넌트를 기반으로 작업하는 방법을 알게 되었고, 프론트엔드 개발자도 백엔드 분야의 기본적인 지식은 알아야한다는 것을 느꼈습니다.",
-    "skills": "HTML, JavaScript, React, Redux, SASS, Bootstrap, Git, GitHub, VS Code, ESLint, Vercel, Figma, Slack, Notion, KakaoTalk",
+    "skills": "React, JavaScript, Redux, SASS, Bootstrap, Git, GitHub, VS Code, ESLint, Vercel, Figma, Slack, Notion, KakaoTalk",
     "response": "1. 날씨 위젯: geolocation API로 사용자의 현재 위치를 확인하여 날씨 정보 출력\n2. 미세먼지 위젯: '서울시 권역별 실시간 대기환경 정보'에서 강동구 정보만 추출하여 출력, 측정결과에 따라 3종의 다른 디자인 지정\n3. Not Found 페이지: 예외처리를 위한 페이지 제작, 디자인, 라우팅 연결",
     "links": [
       {
@@ -15,8 +36,9 @@
     ]
   },
   {
-    "id": 2,
+    "id": 4,
     "title": "Frontend - 반려동물 관련 공간 예약 사이트 '루미펫'",
+    "tier": "secondary",
     "image": ["team2-0.png", "team2-1.png", "team2-2.png"],
     "date": "2024.03.13 ~ 03.28 / 5인",
     "descript": "반려동물이 갑자기 아플 때 24시간 병원을 찾기 어려웠다는 실제 팀원의 경험에서 착안하여 기획 및 제작한 사이트입니다. 공간 예약에는 회원정보가 필요하여 로그인, 회원가입 페이지를 담당했습니다. 코딩 팀플을 처음 참여해보며, 하나의 페이지를 디자인부터 기능까지 제대로 만들어내기 위해 많은 양의 CSS, JS 코드가 필요하다는 것을 깨달았습니다. 또한 각자 담당하는 페이지를 만들고, 그 페이지를 모아서 하나의 사이트를 구축해내는 과정에서 스스로의 성장을 느꼈습니다.",
@@ -30,8 +52,9 @@
     ]
   },
   {
-    "id": 3,
+    "id": 5,
     "title": "UI/UX - 'EVERPOWER' 공식 쇼핑몰",
+    "tier": "secondary",
     "image": ["team1-1.png", "team1-2.png"],
     "date": "2024.01.12 ~ 01.28 / 4인",
     "descript": "팀원 모두가 패션을 좋아해서 2030세대를 타겟층으로 하는 가상의 스포츠웨어 브랜드 쇼핑몰 페이지를 기획 및 디자인했습니다. Figma를 사용한 UI/UX 프로젝트였고, 팀장이라 조금 더 많은 페이지를 담당했습니다. 팀원들과 함께 실제 존재하는 사이트처럼 모든 페이지를 제작해보며, 예쁜 UI만큼 고객들이 편하게 사용할 수 있는 UX도 중요하다는 것을 느꼈습니다. 또한 기획은 상황에 따라 수시로 변화할 수 있으니 항상 유연하게 생각해야 한다는 것을 배웠습니다.",
@@ -45,8 +68,9 @@
     ]
   },
   {
-    "id": 4,
+    "id": 6,
     "title": "Frontend - 영화 리뷰 기록 및 공유 사이트 '시네톡'",
+    "tier": "secondary",
     "image": [
       "cinetalk0.png",
       "cinetalk1.png",
@@ -55,8 +79,8 @@
     ],
     "date": "2024.07.01 ~ 08.08 / 1인",
     "descript": "영화를 보고 리뷰를 기록하는 것이 취미여서, 리뷰에 중점을 둔 사이트로 기획 및 제작했습니다. 공부하고 있던 Swiper를 다양하게 활용해볼 수 있었고, 익숙하지 않던 Redux 상태 관리를 구현하며 작동 원리를 확실히 이해하게 되었습니다.",
-    "skills": "HTML, JavaScript, React, Redux, SASS, MUI, Swiper, Git, GitHub, VS Code, ESLint, Prettier, Vercel, Figma",
-    "response": "메인 페이지의 카드 슬라이드(Swiper), 리뷰 확인/삭제/좋아요/공유하기, 로그인/로그아웃 검증(Redux), 영화정보 검색(영화진흥위원회 API), 반응형 미디어쿼리 적용 등",
+    "skills": "React, JavaScript, Redux, SASS, MUI, Swiper, Git, GitHub, VS Code, ESLint, Prettier, Vercel, Figma",
+    "response": "- 메인 페이지의 카드 슬라이드(Swiper)\n- 리뷰 확인/삭제/좋아요/공유하기\n- 로그인/로그아웃 검증(Redux)\n- 영화정보 검색(영화진흥위원회 API) 등",
     "links": [
       {
         "website": "https://cinetalk.vercel.app/",
@@ -65,13 +89,14 @@
     ]
   },
   {
-    "id": 5,
+    "id": 7,
     "title": "Frontend - 개인 포트폴리오 사이트 ver.2",
+    "tier": "primary",
     "image": ["portfolio1.png", "portfolio2.png", "portfolio3.png"],
-    "date": "2024.08.06 ~ 08.10 / 1인",
+    "date": "- 2024.08.06 ~ 08.10 (초기 개발)\n- 2026.02 ~ (TypeScript 마이그레이션)",
     "descript": "개인의 기술 스택과 작업 내역 정리를 위해 기획 및 제작한 포트폴리오 사이트입니다. 코딩 공부 초기에 순수 JS로 만든 사이트를 React와 SASS를 사용해 리뉴얼하며, 반복 사용되는 요소들을 컴포넌트로 관리하는 것이 작업의 효율성과 생산성 측면에서 우수하다고 다시 한번 느꼈습니다.",
-    "skills": "HTML, JavaScript, React, SASS, MUI, GSAP, Swiper, Git, GitHub, VS Code, ESLint, Prettier, Vercel",
-    "response": "'About' 페이지 타이핑 애니메이션과 GSAP Scroll Trigger, Header 호버 언더라인 이벤트와 페이지 라우팅, '프로젝트' 페이지 Image Carousel (Swiper), 반응형 미디어쿼리 적용 등",
+    "skills": "React, TypeScript, JavaScript, SASS, MUI, GSAP, Swiper, Git, GitHub, VS Code, ESLint, Prettier, Vercel",
+    "response": "- 'About' 페이지 타이핑 애니메이션과 GSAP Scroll Trigger\n- Header 호버 언더라인 이벤트와 페이지 라우팅\n- '프로젝트' 페이지 Image Carousel (Swiper) 등",
     "links": [
       {
         "website": "https://hjinn0813.vercel.app",
